pyproject.toml
[project]
name = "fatima"
version = "0.1.0"
description = "Fatima Fellowship 2026 - Base Model Blind Spot Analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "modal>=0.73.0",
    "torch>=2.2.0",
    "transformers @ git+https://github.com/huggingface/transformers.git",
    "accelerate>=0.34.0",
    "datasets>=3.0.0",
    "huggingface_hub>=0.25.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "qwen-vl-utils>=0.0.8",
    "lm-eval>=0.4.6",
    "bitsandbytes>=0.43.0",
    "sentencepiece>=0.2.0",
]
